Replace the package pipeline framework with concrete middleware
The descriptor + factory + type-parameterized pipeline was more machinery
than three endpoints warrant. Spell the three routes out explicitly in
index.ts and give each stage a concrete middleware that reads top to bottom:

  size:          request -> error -> blocklist -> resolve+serve cache
                 -> rate limit -> build
  exports:       request -> error -> blocklist -> resolve -> build
  exports-sizes: request -> error -> blocklist -> resolve -> serve cache
                 -> rate limit -> build

Only small leaf helpers stay shared (cachePolicy, sizeCacheMaxAge, logCache,
serveFailureCache) in results/packageCache.ts. Drops the generic
PackageResultCache, resolveCachedPackage, packageApiPipeline, PackageEndpoint
descriptors, buildAndStore/resolveAndServeCached factories, and the
runCancellableBuild helper (client-disconnect cancellation is inlined). The
size fast path (serve an exact cached version before hitting npm) lives in a
concrete, testable lookupPackageSize; SSR uses readCachedPackageSize.
